feat: extract NavigationSidebar to shared UI folder and reorganize structure
