+++ vite.config.ts
@@ -11,8 +11,10 @@ export default defineConfig({
   plugins: [react()],
   resolve: {
     alias: {
-      "@shared": path.resolve(__dirname, "./src/shared"),
-      "@pages": path.resolve(__dirname, "./src/pages")
+      "@pages": path.resolve(__dirname, "./src/pages"),
+      "@services": path.resolve(__dirname, "./src/shared/services"),
+      "@types": path.resolve(__dirname, "./src/shared/types"),
+      "@UI": path.resolve(__dirname, "./src/shared/UI"),
     }
   }
 })
